// mlflow/server/js/vendor/design-system/src/theme/_generated/SemanticColors-Light.ts
export const lightColorList: Record<ValidSemanticColors, string> = {
  actionDangerDefaultBackgroundDefault: 'rgba(200, 45, 76, 0.0000)',
  actionDangerDefaultBackgroundHover: 'rgba(200, 45, 76, 0.0800)',
  actionDangerDefaultBackgroundPress: 'rgba(200, 45, 76, 0.1600)',
  actionDangerDefaultBorderDefault: primitives.red600,
  actionDangerDefaultBorderHover: primitives.red700,
  actionDangerDefaultBorderPress: primitives.red800,
  actionDangerDefaultTextDefault: primitives.red600,
  actionDangerDefaultTextHover: primitives.red700,
  actionDangerDefaultTextPress: primitives.red800,
  actionDangerPrimaryBackgroundDefault: primitives.red600,
  actionDangerPrimaryBackgroundHover: primitives.red700,
  actionDangerPrimaryBackgroundPress: primitives.red800,
  actionDangerPrimaryText: primitives.white,
  actionDefaultBackgroundDefault: 'rgba(34, 114, 180, 0.0000)',
  actionDefaultBackgroundHover: 'rgba(34, 114, 180, 0.0800)',
  actionDefaultBackgroundPress: 'rgba(34, 114, 180, 0.1600)',
  actionDefaultBorderDefault: primitives.grey300,
  actionDefaultBorderFocus: primitives.blue600,
  actionDefaultBorderHover: primitives.blue600,
  actionDefaultBorderPress: primitives.blue800,
  actionDefaultIconDefault: primitives.grey600,
  actionDefaultIconHover: primitives.blue700,
  actionDefaultIconPress: primitives.blue800,
  actionDefaultTextDefault: primitives.grey800,
  actionDefaultTextHover: primitives.blue700,
  actionDefaultTextPress: primitives.blue800,
  actionDisabledBackground: primitives.grey100,
  actionDisabledBorder: primitives.grey200,
  actionDisabledText: primitives.grey350,
  actionIconBackgroundDefault: 'rgba(34, 114, 180, 0.0000)',
  actionIconBackgroundHover: 'rgba(34, 114, 180, 0.0800)',
  actionIconBackgroundPress: 'rgba(34, 114, 180, 0.1600)',
  actionIconIconDefault: primitives.grey600,
  actionIconIconHover: primitives.blue700,
  actionIconIconPress: primitives.blue800,
  actionLinkDefault: primitives.blue600,
  actionLinkHover: primitives.blue700,
  actionLinkPress: primitives.blue800,
  actionPrimaryBackgroundDefault: primitives.blue600,
  actionPrimaryBackgroundHover: primitives.blue700,
  actionPrimaryBackgroundPress: primitives.blue800,
  actionPrimaryIcon: primitives.white,
  actionPrimaryTextDefault: primitives.white,
  actionPrimaryTextHover: primitives.white,
  actionPrimaryTextPress: primitives.white,
  actionTertiaryBackgroundDefault: 'rgba(34, 114, 180, 0.0000)',
  actionTertiaryBackgroundHover: 'rgba(34, 114, 180, 0.0800)',
  actionTertiaryBackgroundPress: 'rgba(34, 114, 180, 0.1600)',
  actionTertiaryIconDefault: primitives.blue600,
  actionTertiaryIconHover: primitives.blue700,
  actionTertiaryIconPress: primitives.blue800,
  actionTertiaryTextDefault: primitives.blue600,
  actionTertiaryTextHover: primitives.blue700,
  actionTertiaryTextPress: primitives.blue800,
  backgroundDanger: primitives.red100,
  backgroundPrimary: primitives.white,
  backgroundSecondary: primitives.grey050,
  backgroundSuccess: primitives.green100,
  backgroundWarning: primitives.yellow100,
  border: primitives.grey200,
  borderAccessible: primitives.grey500,
  borderDanger: primitives.red300,
  borderWarning: primitives.yellow300,
  codeBackground: 'rgba(68, 83, 95, 0.0800)',
  overlayOverlay: 'rgba(64, 63, 63, 0.7200)',
  progressFill: primitives.grey300,
  progressTrack: primitives.grey100,
  skeleton: 'rgba(144, 164, 181, 0.1600)',
  tableBackgroundSelectedDefault: 'rgba(68, 83, 95, 0.0800)',
  tableBackgroundSelectedHover: 'rgba(68, 83, 95, 0.1200)',
  tableBackgroundUnselectedDefault: primitives.white,
  tableBackgroundUnselectedHover: 'rgba(68, 83, 95, 0.0400)',
  textPlaceholder: primitives.grey500,
  textPrimary: primitives.grey800,
  textSecondary: primitives.grey500,
  textValidationDanger: primitives.red600,
  textValidationSuccess: primitives.green600,
  textValidationWarning: primitives.yellow600,
  tooltipBackgroundTooltip: primitives.grey800,
};